# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FireboxTechStudios API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services (admin)
      parameters:
        - name: published
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Service" }
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
...
  /services/public:
    get:
      operationId: listPublicServices
      tags: [services]
      summary: List all published services (public)
      parameters:
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Published services
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Service" }
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ─── PORTFOLIO ────────────────────────────────────────────────────────────
  /portfolio:
    get:
      operationId: listProjects
      tags: [portfolio]
      summary: List all projects (admin)
      responses:
        "200":
          description: Projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
    post:
      operationId: createProject
      tags: [portfolio]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project" 
...
  /portfolio/public:
    get:
      operationId: listPublicProjects
      tags: [portfolio]
      summary: List all published projects (public)
      parameters:
        - name: technology
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Published projects
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Project" }
  /portfolio/{id}:
    get:
      operationId: getProject
      tags: [portfolio]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [portfolio]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [portfolio]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ─── TUTORIALS ────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List all reviews (admin)
      parameters:
        - name: status
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Review" }
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update review status (approve/reject/feature/reply)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ─── CONTACT MESSAGES ─────────────────────────────────────────────────────
...
  /quotes:
    get:
      operationId: listQuoteRequests
      tags: [quotes]
      summary: List all quote requests (admin)
      responses:
        "200":
          description: Quote requests
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/QuoteRequest" }
    post:
      operationId: createQuoteRequest
      tags: [quotes]
      summary: Submit a quote request (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
  /quotes/{id}:
    patch:
      operationId: updateQuoteRequest
      tags: [quotes]
      summary: Update quote request status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
    delete:
      operationId: deleteQuoteRequest
      tags: [quotes]
      summary: Delete a quote request
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ─── NEWSLETTER ───────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, title, description, category, published, featured, createdAt]
      properties:
        id: { type: number }
        title: { type: string }
        description: { type: string }
        category: { type: string }
        icon: { type: ["string", "null"] }
        bannerUrl: { type: ["string", "null"] }
        galleryUrls: { type: ["string", "null"] }
        pricing: { type: ["string", "null"] }
        buttonText: { type: ["string", "null"] }
        destinationUrl: { type: ["string", "null"] }
        published: { type: boolean }
        featured: { type: boolean }
        sortOrder: { type: number }
        createdAt: { type: string }
        updatedAt: { type: string }
    ServiceInput:
      type: object
      required: [title, description, category]
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        icon: { type: string }
        bannerUrl: { type: string }
        galleryUrls: { type: string }
        pricing: { type: string }
        buttonText: { type: string }
        destinationUrl: { type: string }
        published: { type: boolean }
        featured: { type: boolean }
        sortOrder: { type: number }
    ServiceUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        icon: { type: string }
        bannerUrl: { type: string }
        galleryUrls: { type: string }
        pricing: { type: string }
        buttonText: { type: string }
        destinationUrl: { type: string }
        published: { type: boolean }
        featured: { type: boolean }
        sortOrder: { type: number }

    # ─── PROJECT ────────────────────────────────────────────────────────────
    Project:
      type: object
      required: [id, title, description, published, featured, createdAt]
      properties:
        id: { type: number }
        title: { type: string }
        description: { type: string }
        technologies: { type: ["string", "null"] }
        screenshotUrls: { type: ["string", "null"] }
        videoUrl: { type: ["string", "null"] }
        githubUrl: { type: ["string", "null"] }
        liveDemoUrl: { type: ["string", "null"] }
        published: { type: boolean }
        featured: { type: boolean }
        category: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProjectInput:
      type: object
      required: [title, description]
      properties:
        title: { type: string }
        description: { type: string }
        technologies: { type: string }
        screenshotUrls: { type: string }
        videoUrl: { type: string }
        githubUrl: { type: string }
        liveDemoUrl: { type: string }
        published: { type: boolean }
        featured: { type: boolean }
        category: { type: string }
    ProjectUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        technologies: { type: string }
        screenshotUrls: { type: string }
        videoUrl: { type: string }
        githubUrl: { type: string }
        liveDemoUrl: { type: string }
        published: { type: boolean }
        featured: { type: boolean }
        category: { type: string }

    # ─── TUTORIAL ───────────────────────────────────────────────────────────
...
    Review:
      type: object
      required: [id, name, email, rating, testimonial, status, featured, createdAt]
      properties:
        id: { type: number }
        name: { type: string }
        email: { type: string }
        rating: { type: number }
        testimonial: { type: string }
        status: { type: string }
        featured: { type: boolean }
        adminReply: { type: ["string", "null"] }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [name, email, rating, testimonial]
      properties:
        name: { type: string }
        email: { type: string }
        rating: { type: number }
        testimonial: { type: string }
    ReviewUpdate:
      type: object
      properties:
        status: { type: string }
        featured: { type: boolean }
        adminReply: { type: string }
...
    QuoteRequest:
      type: object
      required: [id, name, email, projectType, description, budget, status, createdAt]
      properties:
        id: { type: number }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        projectType: { type: string }
        description: { type: string }
        budget: { type: ["string", "null"] }
        timeline: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
    QuoteRequestInput:
      type: object
      required: [name, email, projectType, description]
      properties:
        name: { type: string }
        email: { type: string }
        phone: { type: string }
        projectType: { type: string }
        description: { type: string }
        budget: { type: string }
        timeline: { type: string }
    QuoteRequestUpdate:
      type: object
      properties:
        status: { type: string }

    # ─── NEWSLETTER ─────────────────────────────────────────────────────────
...