# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * FireboxTechStudios API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  galleryUrls?: string | null;
  /** @nullable */
  pricing?: string | null;
  /** @nullable */
  buttonText?: string | null;
  /** @nullable */
  destinationUrl?: string | null;
  published: boolean;
  featured: boolean;
  sortOrder?: number;
  createdAt: string;
  updatedAt?: string;
}